Changes to Upload Script
